+++ src/settings/MappingProfiles/MappingProfilesTransformationsModal/MappingProfilesTransformationsModal.js
@@ -1,32 +1,34 @@
 import React, {
   useState,
+  useCallback,
   useEffect,
 } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { noop } from 'lodash';
+import {
+  get,
+  noop,
+} from 'lodash';
 
 import {
   Button,
-  Icon,
   Modal,
   Pane,
   Paneset,
-  SearchField,
-  AccordionSet,
-  Accordion,
-  FilterAccordionHeader,
-  Checkbox,
 } from '@folio/stripes/components';
 import {
   ExpandFilterPaneButton,
   CollapseFilterPaneButton,
 } from '@folio/stripes/smart-components';
 
+import { recordTypes } from '../RecordTypeField';
 import { TransformationsForm } from './TransformationsForm';
-import { mappingProfileTransformations } from './TransformationsField/transformations';
+import { SearchForm } from './SearchForm';
 
 import css from './MappingProfilesTransformationsModal.css';
+
+const initialSearchFormValues = { filters: { recordTypes: recordTypes.map(recordType => recordType.value) } };
+const fullWidthStyle = { style: { width: '100%' } };
 
 export const MappingProfilesTransformationsModal = ({
   isOpen,
@@ -34,19 +36,26 @@ export const MappingProfilesTransformationsModal = ({
   onCancel,
 }) => {
   const [isFilterPaneOpen, setFilterPaneOpen] = useState(true);
+  const [searchFilters, setSearchFilters] = useState(initialSearchFormValues.filters);
 
   const totalSelectedCount = 0;
-  const searchResultsCount = mappingProfileTransformations.length;
 
   useEffect(() => {
     if (!isOpen) {
       setFilterPaneOpen(true);
+      setSearchFilters(initialSearchFormValues.filters);
     }
   }, [isOpen]);
 
-  const toggleFilterPane = () => {
-    setFilterPaneOpen(curState => !curState);
-  };
+  const searchResults = initialTransformationsValues.transformations
+    .filter(value => get(searchFilters, 'recordTypes', []).includes(value.recordType));
+
+  const toggleFilterPane = useCallback(() => setFilterPaneOpen(curState => !curState), []);
+
+  const handleFiltersChange = useCallback((key, value) => setSearchFilters(curFilters => ({
+    ...curFilters,
+    [key]: value,
+  })), []);
 
   const renderFooter = () => {
     return (
@@ -89,63 +98,19 @@ export const MappingProfilesTransformationsModal = ({
       onClose={onCancel}
     >
       <Paneset>
-        {isFilterPaneOpen && (
-          <Pane
-            data-test-transformations-search-pane
-            defaultWidth="30%"
-            paneTitle={<FormattedMessage id="ui-data-export.searchAndFilter" />}
-            lastMenu={<CollapseFilterPaneButton onClick={toggleFilterPane} />}
-          >
-            <form>
-              <div className={css.searchGroupWrap}>
-                <FormattedMessage id="ui-data-export.mappingProfiles.transformations.searchFields">
-                  {label => (
-                    <SearchField
-                      data-test-transformations-search-field
-                      aria-label={label}
-                      autoFocus
-                      marginBottom0
-                    />
-                  )}
-                </FormattedMessage>
-              </div>
-              <div>
-                <Button
-                  data-test-transformations-reset
-                  buttonStyle="none"
-                >
-                  <Icon icon="times-circle-solid">
-                    <FormattedMessage id="ui-data-export.resetAll" />
-                  </Icon>
-                </Button>
-              </div>
-              <AccordionSet id="transformations-filter-accordions">
-                <Accordion
-                  id="transformations-record-type-accordion"
-                  header={FilterAccordionHeader}
-                  label={<FormattedMessage id="ui-data-export.recordType" />}
-                  separator={false}
-                >
-                  <Checkbox
-                    label={<FormattedMessage id="stripes-data-transfer-components.recordTypes.instance" />}
-                    fullWidth
-                    checked
-                  />
-                  <Checkbox
-                    label={<FormattedMessage id="stripes-data-transfer-components.recordTypes.holdings" />}
-                    fullWidth
-                    checked
-                  />
-                  <Checkbox
-                    label={<FormattedMessage id="stripes-data-transfer-components.recordTypes.item" />}
-                    fullWidth
-                    checked
-                  />
-                </Accordion>
-              </AccordionSet>
-            </form>
-          </Pane>
-        )}
+        <Pane
+          data-test-transformations-search-pane
+          defaultWidth="30%"
+          paneTitle={<FormattedMessage id="ui-data-export.searchAndFilter" />}
+          lastMenu={<CollapseFilterPaneButton onClick={toggleFilterPane} />}
+          hidden={!isFilterPaneOpen}
+        >
+          <SearchForm
+            initialValues={initialSearchFormValues}
+            onFiltersChange={handleFiltersChange}
+            onSubmit={noop}
+          />
+        </Pane>
         <Pane
           data-test-transformations-results-pane
           defaultWidth="fill"
@@ -154,13 +119,15 @@ export const MappingProfilesTransformationsModal = ({
           paneSub={(
             <FormattedMessage
               id="ui-data-export.mappingProfiles.transformations.searchResultsCountHeader"
-              values={{ count: searchResultsCount }}
+              values={{ count: searchResults.length }}
             />
           )}
           firstMenu={!isFilterPaneOpen ? <ExpandFilterPaneButton onClick={toggleFilterPane} /> : null}
+          {...(!isFilterPaneOpen && fullWidthStyle)}
         >
           <TransformationsForm
             initialValues={initialTransformationsValues}
+            searchResults={searchResults}
             autosize
             onSubmit={noop}
           />