+++ src/settings/MappingProfiles/MappingProfilesTransformationsModal/TransformationsForm/TransformationsForm.js
@@ -1,5 +1,9 @@
-import React from 'react';
+import React, {
+  memo,
+  useCallback,
+} from 'react';
 import PropTypes from 'prop-types';
+import { intersectionWith } from 'lodash';
 
 import stripesFinalForm from '@folio/stripes/final-form';
 
@@ -7,21 +11,31 @@ import { TransformationField } from '../TransformationsField';
 
 import css from './TransformationsForm.css';
 
-const TransformationsFormComponent = ({
+const TransformationsFormComponent = memo(({
   autosize,
+  searchResults,
   handleSubmit,
 }) => {
+  const prepareData = useCallback(
+    results => intersectionWith(results, searchResults, (val1, val2) => val1.path === val2.path),
+    [searchResults],
+  );
+
   return (
     <form
       className={css.transformationsForm}
       onSubmit={handleSubmit}
     >
-      <TransformationField autosize={autosize} />
+      <TransformationField
+        autosize={autosize}
+        prepareData={prepareData}
+      />
     </form>
   );
-};
+});
 
 TransformationsFormComponent.propTypes = {
+  searchResults: PropTypes.arrayOf(PropTypes.object).isRequired,
   handleSubmit: PropTypes.func.isRequired,
   autosize: PropTypes.bool,
 };