+++ test/bigtest/tests/units/settings/MappingProfilesForm/interactors/TransformationsModalInteractor.js
@@ -8,19 +8,21 @@ import ButtonInteractor from '@folio/stripes-components/lib/Button/tests/interac
 import CheckboxInteractor from '@folio/stripes-components/lib/Checkbox/tests/interactor';
 import TextFieldInteractor from '@folio/stripes-components/lib/TextField/tests/interactor';
 import { AccordionSetInteractor } from '@folio/stripes-components/lib/Accordion/tests/interactor';
+
+import { PaneInteractor } from '../../../../../interactors';
 import { TransformationsInteractor } from './TransformationsInteractor';
 
 @interactor
 export class TransformationsModalInteractor {
   static defaultScope = '[data-test-transformations-modal]';
 
-  searchPane = new Interactor('[data-test-transformations-search-pane]');
+  searchPane = new PaneInteractor('[data-test-transformations-search-pane]');
   collapseSearchPaneButton = new Interactor('[data-test-collapse-filter-pane-button]');
   expandSearchPaneButton = new Interactor('[data-test-expand-filter-pane-button]');
-  resultsPane = new Interactor('[data-test-transformations-results-pane]');
+  resultsPane = new PaneInteractor('[data-test-transformations-results-pane]');
   filterAccordions = new AccordionSetInteractor('#transformations-filter-accordions');
   searchField = new TextFieldInteractor('[data-test-transformations-search-field]');
-  recordTypeCheckboxes = collection('#transformations-record-type-accordion [data-test-checkbox]', CheckboxInteractor);
+  recordTypeFilters = collection('#transformations-record-type-accordion [data-test-checkbox]', CheckboxInteractor);
   totalSelected = new Interactor('[data-test-transformations-total-selected]');
   saveButton = new ButtonInteractor('[data-test-transformations-save]');
   resetButton = new ButtonInteractor('[data-test-transformations-reset]');